Final debug

+++ libft_repo/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/06 17:15:24 by wlin              #+#    #+#             */
-/*   Updated: 2017/06/14 15:27:36 by wlin             ###   ########.fr       */
+/*   Updated: 2017/06/17 14:52:07 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(char *str)
+size_t		ft_strlen(char *str)
 {
 	char *str_ptr;
